fix: encoding for string values with emojis
Currently the encoding and decoding a value like `["Get takeaway now 🥡!"]` will fail during the decoding step. This is because the length prefix is larger than the actual length of the string and so parsing the string swallows the end of list symbol.

I think using the text encoder is not necessary and we can just use the string length as when reding the data we will be just selecting the substring from the encoded value. Though there might be edge cases I'm missing.

I've added test cases to test for correct parsing of strings with emojis, but I think it would also be useful to have a test that encodes and decodes data and checks that the output matches the input. Something like this would catch this issue if there was a test case with the example similar to the one above. Though this is probably something that can be done in another PR.